move battle data
