+++ src/battleEngine/data/initialize.js
@@ -3,7 +3,7 @@
 const { registerMoves, registerLegacyMoves } = require("./moveService");
 const { registerEffects, registerLegacyEffects } = require("./effectRegistry");
 const { effectsToRegister } = require("./effects");
-const { moveConfig, moveExecutes, effectConfig } = require("./battleConfig");
+const { moveConfig, moveExecutes, effectConfig } = require("../battleConfig");
 const { movesToRegister } = require("./moves");
 
 const initialize = () => {
