+++ src/services/mythic.js
@@ -46,7 +46,7 @@ const {
   calculateAndUpdatePokemonStats,
 } = require("./pokemon");
 const { getTrainer, updateTrainer } = require("./trainer");
-const { getMoves } = require("../battleEngine/moveService");
+const { getMoves } = require("../battleEngine/data/moveService");
 
 const getMythic = async (trainer, speciesId) => {
   const speciesData = pokemonConfig[speciesId];
